+++ bot.py
@@ -1,6 +1,7 @@
 import os
 
 from dotenv import load_dotenv
+import botctypes
 import datetime
 import discord
 from discord.ext import commands, tasks
@@ -11,6 +12,8 @@ from pymongo import MongoClient
 import random
 import shlex
 import traceback
+
+import votetimer
 
 load_dotenv()
 TOKEN = os.getenv('DISCORD_TOKEN')
@@ -36,6 +39,8 @@ if MONGO_CONNECT is None:
 cluster = MongoClient(MONGO_CONNECT)
 db = cluster[MONGO_DB]
 
+COMMAND_PREFIX = os.getenv('COMMAND_PREFIX') or '!'
+
 g_dbGuildInfo = db['GuildInfo']
 g_dbActiveGames = db['ActiveGames']
 
@@ -57,7 +62,7 @@ def getCategory(guild, name, catId):
 # Get a channel by ID or name, preferring ID
 def getChannelFromCategory(category, name, chanId):
     chanById = discord.utils.find(lambda c: (c.type == discord.ChannelType.voice or c.type == discord.ChannelType.text) and c.id == chanId, category.channels)
-    return chanById or  discord.utils.find(lambda c: (c.type == discord.ChannelType.voice or c.type == discord.ChannelType.text) and c.name == name, category.channels)
+    return chanById or discord.utils.find(lambda c: (c.type == discord.ChannelType.voice or c.type == discord.ChannelType.text) and c.name == name, category.channels)
 
 # Get a role by ID or name, preferring ID
 def getRole(guild, name, roleId):
@@ -105,6 +110,10 @@ class TownInfo:
             self.storyTellerRole = getRole(guild, document["storyTellerRole"], document["storyTellerRoleId"])
             self.villagerRole = getRole(guild, document["villagerRole"], document["villagerRoleId"])
 
+            self.chatChannel = None
+            if 'chatChannel' in document and 'chatChannelId' in document:
+                self.chatChannel = getChannelFromCategory(self.dayCategory, document['chatChannel'], document['chatChannelId'])
+
             activePlayers = set()
             for c in self.dayChannels:
                 activePlayers.update(c.members)
@@ -128,12 +137,20 @@ class botcBot(commands.Bot):
 
     # Get a well-defined TownInfo based on the stored DB info for this guild
     def getTownInfo(self, ctx):
+        return self.getTownInfoByIds(ctx.guild.id, ctx.channel.id, ctx.guild)
 
-        query = { "guild" : ctx.guild.id, "controlChannelId" : ctx.channel.id }
+    def getTownInfoByTownId(self, town_id, guild=None):
+        return self.getTownInfoByIds(town_id.guild_id, town_id.channel_id, guild)
+
+    def getTownInfoByIds(self, guild_id, channel_id, guild=None):
+        query = { "guild" : guild_id, "controlChannelId" : channel_id }
         doc = g_dbGuildInfo.find_one(query)
+        
+        if not guild:
+            guild = self.get_guild(guild_id)
 
-        if doc:
-            return TownInfo(ctx.guild, doc)
+        if doc and guild:
+            return TownInfo(guild, doc)
         else:
             return None
     
@@ -148,10 +165,10 @@ class botcBot(commands.Bot):
 
 
     async def on_ready(self):
-        print(f'{self.user.name} has connected to Discord!')
+        print(f'{self.user.name} has connected to Discord! Command prefix: {COMMAND_PREFIX}')
 
 # Setup cog
-class Setup(commands.Cog):
+class SetupCog(commands.Cog, name='Setup'):
     def __init__(self, bot):
         self.bot = bot
 
@@ -187,7 +204,7 @@ class Setup(commands.Cog):
 
         await self.sendEmbed(ctx, info)
 
-    @commands.command(name='addTown', aliases=['addtown'], help='Add a game on this server.\n\nUsage: !addTown <control channel> <town square channel> <day category> <night category> <storyteller role> <villager role>\n\nAlternate usage: !addTown control=<control channel> townSquare=<town square channel> dayCategory=<day category> nightCategory=<night category> stRole=<storyteller role> villagerRole=<villager role>')
+    @commands.command(name='addTown', aliases=['addtown'], help=f'Add a game on this server.\n\nUsage: {COMMAND_PREFIX}addTown <control channel> <town square channel> <day category> <night category> <storyteller role> <villager role> [chat channel]\n\nAlternate usage: {COMMAND_PREFIX}addTown control=<control channel> townSquare=<town square channel> dayCategory=<day category> [nightCategory=<night category>] stRole=<storyteller role> villagerRole=<villager role> [chatChannel=<chat channel>]')
     async def addTown(self, ctx):
         params = shlex.split(ctx.message.content)
 
@@ -204,7 +221,7 @@ class Setup(commands.Cog):
         guild = ctx.guild
 
         params = shlex.split(ctx.message.content)
-        usageStr = "Usage: `!removeTown <day category name>` or `!removeTown` alone if run from the town's control channel"
+        usageStr = f'Usage: `{COMMAND_PREFIX}removeTown <day category name>` or `{COMMAND_PREFIX}removeTown` alone if run from the town\'s control channel'
 
         if len(params) == 1:
             post = {"guild" : guild.id, "controlChannelId" : ctx.channel.id }
@@ -225,6 +242,42 @@ class Setup(commands.Cog):
         else:
             await ctx.send(f"Couldn't find a town to remove! {usageStr}")
 
+    @commands.command(name='setChatChannel', help=f'Set the chat channel associated with this town.\n\nUsage: {COMMAND_PREFIX}setChatChannel <chat channel>')
+    async def set_chat_channel(self, ctx):
+        params = shlex.split(ctx.message.content)
+        
+        if len(params) != 2:
+            await ctx.send(f'Incorrect usage for `{COMMAND_PREFIX}setChatChannel`: should provide `<chat channel>`')
+            return None
+
+        chat_channel_name = params[1]
+        
+        query = { "guild" : ctx.guild.id, "controlChannelId" : ctx.channel.id }
+        doc = g_dbGuildInfo.find_one(query)
+        if not doc:
+            await ctx.send(f'Could not find a town! Are you running this command from the town control channel?')
+            return None
+
+        day_category = getCategory(ctx.guild, doc['dayCategory'], doc['dayCategory'])
+        if not day_category:
+            await ctx.send(f'Could not find the category {doc["dayCategory"]}!')
+            return None
+
+        chat_channel = getChannelFromCategoryByName(day_category, chat_channel_name)
+        if not chat_channel:
+            await ctx.send(f'Could not find the channel {chat_channel_name} in the category {doc["dayCategory"]}!')
+            return None
+
+        doc['chatChannel'] = chat_channel.name
+        doc['chatChannelId'] = chat_channel.id
+        g_dbGuildInfo.replace_one(query, doc, True)
+
+        info = self.bot.getTownInfo(ctx)
+        if info:
+            await self.sendEmbed(ctx, info)
+        else:
+            await ctx.send(f'There was a problem adding {chat_channel_name} to the town.')
+
 
     # Parse all the params for addTown, sanity check them, and return useful dicts
     async def resolveTownInfoParams(self, ctx, params):
@@ -234,6 +287,7 @@ class Setup(commands.Cog):
         nightCatName = None
         stRoleName = None
         villagerName = None
+        chatChannelName = None
 
         hasNamedArgs = False
 
@@ -259,13 +313,15 @@ class Setup(commands.Cog):
                 stRoleName = v
             elif p == "villagerrole":
                 villagerName = v
+            elif p == "chatchannel":
+                chatChannelName = v
             else:
-                await ctx.send(f'Unknown param to `!addTown`: \"{p}\". Valid params: control, townSquare, dayCategory, nightCategory, stRole, villagerRole')
+                await ctx.send(f'Unknown param to `{COMMAND_PREFIX}addTown`: \"{p}\". Valid params: control, townSquare, dayCategory, nightCategory, stRole, villagerRole, chatChannel')
                 return (None, None)
 
         if not hasNamedArgs:
             if len(params) < 7:
-                await ctx.send("Too few params to `!addTown`: should provide `<control channel> <townsquare channel> <day category> <night category> <storyteller role> <villager role>`")
+                await ctx.send(f'Too few params to `{COMMAND_PREFIX}addTown`: should provide `<control channel> <townsquare channel> <day category> <night category> <storyteller role> <villager role> [chat channel]`')
                 return (None, None)
 
             controlName = params[1]
@@ -274,12 +330,14 @@ class Setup(commands.Cog):
             nightCatName = params[4]
             stRoleName = params[5]
             villagerName = params[6]
+            if len(params) > 7:
+                chatChannelName = params[7]
         
-        return await self.resolveTownInfo(ctx, controlName, townSquareName, dayCatName, nightCatName, stRoleName, villagerName)
+        return await self.resolveTownInfo(ctx, controlName, townSquareName, dayCatName, nightCatName, stRoleName, villagerName, chatChannelName)
 
 
     # Using passed-in name params, resolve town info and find all the stuff needed to post to DB
-    async def resolveTownInfo(self, ctx, controlName, townSquareName, dayCatName, nightCatName, stRoleName, villagerName):
+    async def resolveTownInfo(self, ctx, controlName, townSquareName, dayCatName, nightCatName, stRoleName, villagerName, chatChannelName):
         guild = ctx.guild
         
         dayCat = getCategoryByName(guild, dayCatName)
@@ -290,6 +348,10 @@ class Setup(commands.Cog):
 
         stRole = getRoleByName(guild, stRoleName)
         villagerRole = getRoleByName(guild, villagerName)
+
+        chatChannel = None
+        if chatChannelName:
+            chatChannel = getChannelFromCategoryByName(dayCat, chatChannelName)
 
         # TODO: more checks
         # does the night category contain some channels to distribute people to?
@@ -306,6 +368,10 @@ class Setup(commands.Cog):
             await ctx.send(f'Couldn\'t find a channel named `{controlName}` in category `{dayCatName}`!')
             return None
 
+        if chatChannelName and not chatChannel:
+            await ctx.send(f'Couldn\'t find a chat channel named `{chatChannelName}` in category `{dayCatName}`!')
+            return None
+
         if not townSquare:
             await ctx.send(f'Couldn\'t find a channel named `{townSquareName}` in category `{dayCatName}`!')
             return None
@@ -318,14 +384,17 @@ class Setup(commands.Cog):
             await ctx.send(f'Couldn\'t find a role named `{villagerName}`!')
             return None
 
-        # Night category is optional
+        # Some categories/channels are optional
         nightCatId = nightCat and nightCat.id or None
+        chatChannelId = chatChannel and chatChannel.id or None
 
         # Object suitable for mongo
         post = {   
             "guild" : guild.id,
             "controlChannel" : controlName,
             "controlChannelId" : controlChan.id,
+            "chatChannel" : chatChannelName,
+            "chatChannelId" : chatChannelId,
             "townSquare" : townSquareName,
             "townSquareId" : townSquare.id,
             "dayCategory" : dayCatName,
@@ -346,16 +415,16 @@ class Setup(commands.Cog):
         return (post, objs)
 
 
-    @commands.command(name='createTown', aliases=['createtown'], help='Create an entire town on this server, including categories, roles, channels, and permissions.\n\nUsage: !createTown <town name> [server storyteller role] [server player role] [noNight]')
+    @commands.command(name='createTown', aliases=['createtown'], help=f'Create an entire town on this server, including categories, roles, channels, and permissions.\n\nUsage: {COMMAND_PREFIX}createTown <town name> [server storyteller role] [server player role] [noNight]')
     async def createTown(self, ctx):
         params = shlex.split(ctx.message.content)
 
         guild = ctx.guild
         
-        usageStr = "Usage: `!createTown <town name> [server storyteller role] [server player role] [noNight]`"
+        usageStr = f"Usage: `{COMMAND_PREFIX}createTown <town name> [server storyteller role] [server player role] [noNight]`"
 
         if len(params) < 2:
-            await ctx.send("Too few params to `!createTown`. " + usageStr)
+            await ctx.send(f"Too few params to `{COMMAND_PREFIX}createTown`. " + usageStr)
             return None
 
         townName = params[1]
@@ -464,6 +533,7 @@ class Setup(commands.Cog):
             chatChannel = getChannelFromCategoryByName(dayCat, chatChannelName)
             if not chatChannel:
                 chatChannel = await dayCat.create_text_channel(chatChannelName)
+            await chatChannel.set_permissions(botRole, view_channel=True)
             if not guildPlayerRole:
                 await chatChannel.set_permissions(everyoneRole, view_channel=False)
 
@@ -498,8 +568,8 @@ class Setup(commands.Cog):
                         await nightCat.create_voice_channel(nightChannelName)
 
 
-            # Calling !addTown
-            (post, info) = await self.resolveTownInfo(ctx, moverChannelName, townSquareChannelName, dayCatName, nightCatName, gameStRoleName, gameVillagerRoleName)
+            # Calling addTown
+            (post, info) = await self.resolveTownInfo(ctx, moverChannelName, townSquareChannelName, dayCatName, nightCatName, gameStRoleName, gameVillagerRoleName, chatChannelName)
 
             if not post:
                 await ctx.send("There was a problem creating the town of **" + townName + "**.")
@@ -521,7 +591,7 @@ class Setup(commands.Cog):
         usageStr = "Usage: `<town name>`"
 
         if len(params) < 2:
-            await ctx.send("Too few params to `!destroyTown`. " + usageStr)
+            await ctx.send(f"Too few params to `{COMMAND_PREFIX}destroyTown`. " + usageStr)
             return None
 
         townName = params[1]
@@ -663,6 +733,7 @@ class Setup(commands.Cog):
         embed = discord.Embed(title=f'{guild.name} // {townInfo.dayCategory.name}', description=f'Created {townInfo.timestamp} by {townInfo.authorName}', color=0xcc0000)
         embed.add_field(name="Control Channel", value=townInfo.controlChannel.name, inline=False)
         embed.add_field(name="Town Square", value=townInfo.townSquare.name, inline=False)
+        embed.add_field(name="Chat Channel", value=townInfo.chatChannel and townInfo.chatChannel.name or "<None>", inline=False)
         embed.add_field(name="Day Category", value=townInfo.dayCategory.name, inline=False)
         embed.add_field(name="Night Category", value=townInfo.nightCategory and townInfo.nightCategory.name or "<None>", inline=False)
         embed.add_field(name="Storyteller Role", value=townInfo.storyTellerRole.name, inline=False)
@@ -670,9 +741,10 @@ class Setup(commands.Cog):
         await ctx.send(embed=embed)
 
 
-class Gameplay(commands.Cog):
+class GameplayCog(commands.Cog, name='Gameplay'):
     def __init__(self, bot):
         self.bot = bot
+        self.votetimer = votetimer.VoteTimer(bot, self.move_users_for_vote)
 
         # Start the timer if we have active games
         if g_dbActiveGames.count_documents({}) > 0:
@@ -840,7 +912,7 @@ class Gameplay(commands.Cog):
             # grant the storyteller the Current Storyteller role if necessary
             storyTeller = ctx.message.author
             if storyTeller not in info.storyTellers:
-                await ctx.send(f"New storyteller: **{self.getUserName(storyTeller)}**. (Use `!setStorytellers` for 2+ storytellers)")
+                await ctx.send(f"New storyteller: **{self.getUserName(storyTeller)}**. (Use `{COMMAND_PREFIX}setStorytellers` for 2+ storytellers)")
                 await self.setStorytellersInternal(ctx, [storyTeller])
 
             # find additions and deletions by diffing the sets
@@ -894,7 +966,7 @@ class Gameplay(commands.Cog):
         
         return None
 
-    # Common code for parsing !evil and !lunatic commands
+    # Common code for parsing evil and lunatic commands
     async def processMinionMessage(self, ctx, users):
 
         # Split the message allowing quoted substrings
@@ -940,8 +1012,8 @@ class Gameplay(commands.Cog):
         await demonUser.send(demonMsg)
 
     # Command to send fake evil info to the Lunatic
-    # Works the same as !evil, but doesn't message the minions
-    @commands.command(name='lunatic', help='Send fake evil info to the Lunatic. Format is `!lunatic <Lunatic> <fake minion> <fake minion> <fake minion>`')
+    # Works the same as evil, but doesn't message the minions
+    @commands.command(name='lunatic', help=f'Send fake evil info to the Lunatic. Format is `{COMMAND_PREFIX}lunatic <Lunatic> <fake minion> <fake minion> <fake minion>`')
     async def onLunatic(self, ctx):
         if not await self.isValid(ctx):
             return
@@ -960,7 +1032,7 @@ class Gameplay(commands.Cog):
             await self.bot.sendErrorToAuthor(ctx)
 
     # Command to send demon/minion info to the Demon and Minions
-    @commands.command(name='evil', help='Send evil info to evil team. Format is `!evil <demon> <minion> <minion> <minion>`')
+    @commands.command(name='evil', help=f'Send evil info to evil team. Format is `{COMMAND_PREFIX}evil <demon> <minion> <minion> <minion>`')
     async def onEvil(self, ctx):
         if not await self.isValid(ctx):
             return
@@ -1007,7 +1079,7 @@ class Gameplay(commands.Cog):
             info = self.bot.getTownInfo(ctx)
             
             if not info.nightCategory:
-                await ctx.send(f'This town does not have a Night category and therefore does not support the `!day` or `!night` commands. If you want to change this, please add a Night category and use the `!addTown` command to update the town info!')
+                await ctx.send(f'This town does not have a Night category and therefore does not support the `{COMMAND_PREFIX}day` or `{COMMAND_PREFIX}night` commands. If you want to change this, please add a Night category and use the `!addTown` command to update the town info!')
                 return
 
             # get list of users in town square   
@@ -1050,7 +1122,7 @@ class Gameplay(commands.Cog):
             info = self.bot.getTownInfo(ctx)
 
             if not info.nightCategory:
-                await ctx.send(f'This town does not have a Night category and therefore does not support the `!day` or `!night` commands. If you want to change this, please add a Night category and use the `!addTown` command to update the town info!')
+                await ctx.send(f'This town does not have a Night category and therefore does not support the `{COMMAND_PREFIX}day` or `{COMMAND_PREFIX}night` commands. If you want to change this, please add a Night category and use the `!addTown` command to update the town info!')
                 return
 
             # get users in night channels
@@ -1073,6 +1145,20 @@ class Gameplay(commands.Cog):
         except Exception as ex:
             await self.bot.sendErrorToAuthor(ctx)
 
+    async def move_users_for_vote(self, town_info):
+        # get users in day channels other than Town Square
+        users = list()
+        for c in town_info.dayChannels:
+            if c != town_info.townSquare:
+                users.extend(c.members)
+
+        await town_info.controlChannel.send(f'Moving {len(users)} players from daytime channels to **{town_info.townSquare.name}**.')
+
+        # move them to Town Square
+        for user in users:
+            await user.move_to(town_info.townSquare)
+
+
     # Move users from other daytime channels back to Town Square
     @commands.command(name='vote', help='Move players from daytime channels back to Town Square')
     async def onVote(self, ctx):
@@ -1081,18 +1167,7 @@ class Gameplay(commands.Cog):
 
         try:
             info = self.bot.getTownInfo(ctx)
-            
-            # get users in day channels other than Town Square
-            users = list()
-            for c in info.dayChannels:
-                if c != info.townSquare:
-                    users.extend(c.members)
-
-            await ctx.send(f'Moving {len(users)} players from daytime channels to **{info.townSquare.name}**.')
-
-            # move them to Town Square
-            for user in users:
-                await user.move_to(info.townSquare)
+            await self.move_users_for_vote(info)
         
         except Exception as ex:
             await self.bot.sendErrorToAuthor(ctx)
@@ -1160,9 +1235,26 @@ class Gameplay(commands.Cog):
     async def beforecleanupInactiveGames(self):
         await self.bot.wait_until_ready()
 
+    # Start the vote timer
+    @commands.command(name='voteTimer', aliases=['vt', 'votetimer'], help=f'Start a countdown to voting time.\n\nUsage: {COMMAND_PREFIX}votetimer <time string>\n\nTime string can look like: "5 minutes 30 seconds" or "5:30" or "5m30s"')
+    async def start_timer(self, ctx):
+        await self.perform_action_reporting_errors(self.votetimer.start_timer, ctx)
 
-COMMAND_PREFIX = os.getenv('COMMAND_PREFIX') or '!'
+    # Stop an ongoing vote timer
+    @commands.command(name='stopVoteTimer', aliases=['svt', 'stopvotetimer'], help=f'Stop an existing countdown to voting.\n\nUsage: {COMMAND_PREFIX}stopvotetimer')
+    async def stop_timer(self, ctx):
+        await self.perform_action_reporting_errors(self.votetimer.stop_timer, ctx)
+
+    async def perform_action_reporting_errors(self, action, ctx):
+        try:
+            message = await action(ctx)
+            if message != None:
+                await ctx.send(message)
+
+        except Exception as ex:
+            await ctx.bot.sendErrorToAuthor(ctx)
+
 bot = botcBot(command_prefix=COMMAND_PREFIX, intents=intents, description='Bot to manage playing Blood on the Clocktower via Discord')
-bot.add_cog(Setup(bot))
-bot.add_cog(Gameplay(bot))
+bot.add_cog(SetupCog(bot))
+bot.add_cog(GameplayCog(bot))
 bot.run(TOKEN)